Formally document location of meeting  minutes
As there is mainly technical work on VMA done currently, this is the only agenda item. Sometimes small topics come up as a recap from OSS EU or similar which fall under AoB.

I propose to simplify the meeting notes taking by adding this md file for the time being to allow community to better find the working material.